# docs/swagger.yaml
definitions:
  models.Clip:
    properties:
      creator:
        $ref: '#/definitions/models.Creator'
      creator_uuid:
        type: string
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      game:
        $ref: '#/definitions/models.Game'
      game_uuid:
        type: string
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
      viewcount:
        type: integer
      vod:
        $ref: '#/definitions/models.Vod'
      vod_offset:
        type: integer
      vod_uuid:
        type: string
    type: object
  models.Creator:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
  models.Emote:
    properties:
      id:
        type: string
      name:
        type: string
      provider:
        type: string
      url:
        type: string
    type: object
  models.Game:
    properties:
      box_art_url:
        type: string
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
  models.Vod:
    properties:
      clips:
        items:
          $ref: '#/definitions/models.Clip'
        type: array
      date:
        type: string
      duration:
        type: integer
      filename:
        type: string
      fps:
        type: number
      publish:
        type: boolean
      resolution:
        type: string
      size:
        type: integer
      title:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /emotes/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The id of an emote
        in: query
        name: id
        type: string
      - description: The name of an emote
        in: query
        name: name
        type: string
      - description: The provider of an emote
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Emote'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all emotes
      tags:
      - Emotes
    post:
      consumes:
      - application/json
      parameters:
      - description: Emote obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Emote'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create emote
      tags:
      - Emotes
  /emotes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete emote
      tags:
      - Emotes
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: id
        required: true
        type: integer
      - description: Emote obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Emote'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Patch emote
      tags:
      - Emotes
  /games/:
    get:
      consumes:
      - application/json
      parameters:
      - description: The uuid of a game
        in: query
        name: uuid
        type: integer
      - description: The name of a game
        in: query
        name: name
        type: string
      - description: The box_art of a game
        in: query
        name: box_art
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all games
      tags:
      - Games
    post:
      consumes:
      - application/json
      parameters:
      - description: Game obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create game
      tags:
      - Games
  /games/{uuid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete game
      tags:
      - Games
    get:
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Game'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get game by uuid
      tags:
      - Games
    patch:
      consumes:
      - application/json
      parameters:
      - description: Unique Identifier
        in: path
        name: uuid
        required: true
        type: integer
      - description: Game obj
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Patch game
      tags:
      - Games
...
  /stats/long:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get long statistics
      tags:
      - Statistics
  /stats/short:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get short statistics
      tags:
      - Statistics
...
swagger: "2.0"